update ui for tasks

+++ src/app/features/tasks/components/task/task.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, inject, input } from '@angular/core';
 import { take } from 'rxjs';
 import { ButtonComponent } from '../../../shared/components/button/button.component';
-import { Task } from '../../models/task.model';
+import { Task, TaskPriority } from '../../models/task.model';
 import { TaskService } from '../../services/task.service';
 import { GoalsService } from '../../../goals/services/goals.service';
 
@@ -18,6 +18,9 @@ export class TaskComponent {
   taskService = inject(TaskService);
   goalsService = inject(GoalsService);
 
+  // Expose enum to template
+  readonly TaskPriority = TaskPriority;
+
   toggleDescription(taskId: string): void {
     if (this.expandedTaskId === taskId) {
       this.expandedTaskId = null;
@@ -31,15 +34,17 @@ export class TaskComponent {
   }
 
   deleteTask(id: string) {
-    this.taskService.deleteTask(id).subscribe({
-      next: (response) => {
-        if (this.task().completed) {
-          this.taskService.completeTasks.reload();
-        } else {
-          this.taskService.incompleteTasks.reload();
-        }
-      },
-    });
+    if (confirm('Are you sure you want to delete this task?')) {
+      this.taskService.deleteTask(id).subscribe({
+        next: (response) => {
+          if (this.task().completed) {
+            this.taskService.completeTasks.reload();
+          } else {
+            this.taskService.incompleteTasks.reload();
+          }
+        },
+      });
+    }
   }
 
   completeTask(boolean: boolean, event?: Event) {
